* 'feature/product_edit_form' of https://github.com/akeneo/pim-community-dev:
  PIM-4187 : fix PEF behat features
  PIM-4187 Split ProductRestController::postAction() into smaller methods

+++ features/Context/FixturesContext.php
@@ -1535,7 +1535,9 @@ class FixturesContext extends RawMinkContext
         $mediaManager = $this->getMediaManager();
         $allMedia = $product->getMedia();
         foreach ($allMedia as $media) {
-            unlink($mediaManager->getFilePath($media));
+            if ($media) {
+                unlink($mediaManager->getFilePath($media));
+            }
         }
     }
